back_end/requirements.txt
Django==5.0
asgiref==3.7.2
dj_database_url==2.1.0
gunicorn==21.2.0
packaging==23.2
psycopg2-binary==2.9.9
psycopg2==2.9.9
python-dotenv==1.0.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
whitenoise==6.6.0